SSCSCI-969: fix failing tests
