+++ src/contractTest/java/uk/gov/hmcts/reform/sscs/evidenceshare/SendLetterServiceConsumerTest.java
@@ -7,7 +7,7 @@ import au.com.dius.pact.consumer.junit5.PactTestFor;
 import au.com.dius.pact.core.model.PactSpecVersion;
 import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
-import au.com.dius.pact.core.model.annotations.PactFolder;
+import au.com.dius.pact.core.model.annotations.PactDirectory;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -38,7 +38,7 @@ import uk.gov.hmcts.reform.sendletter.api.proxy.SendLetterApiProxy;
 @ExtendWith(SpringExtension.class)
 @ExtendWith(PactConsumerTestExt.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@PactFolder("pacts")
+@PactDirectory("pacts")
 @PactTestFor(providerName = "rpeSendLetterService_SendLetterController", port = "4021")
 @TestPropertySource(locations = {"classpath:application.properties"})
 public class SendLetterServiceConsumerTest {
